Update MainWindow.xaml.cs
Try to handle file access issues. Other access issues will try to try catch so we may as well here and sleep it if needed. This happens in cases where something gets extracted and the extracted thing is extractable and touched on right as you're trying to write to it again.